Changed distance implementation and handling of zeros in SLRs
stats::dist worked well for manhattan, euclidean, and maximum distances, but doesn't calculate absolute or cosine distances. All 5 distances can be calculated with outer.

Previously, NAs were set to 0.01 in denominator scores. But zeros were not changed, which would result in an error. Now NAs and 0 in the denominator are set to parameter zero_correction (default 1e-10).