Moved processing steps inside `calculate_slr`
Moved processing steps inside `calculate_slr` so the user only needs to input paths for 2 handwriting samples. Optionally, the user can also set a project directory. The use also has the option to copy the samples to the project directory.

By default the project directory is NULL. If the user does not specify a project directory, the helper files are saved to a directory in the temp directory which is deleted at the end of the function.

`calculate_slr` now returns an error message if the two input paths are identical. This check was added because if the two paths are identical, only one output file is created in the graphs folder and the clusters folder because the "two" input documents have the same file name. Then `get_distances` throws an error.