Fixed bug in `absolute_dist` when cluster fill rates for a cluster are all zero.
The `outer` function called by the helper function `absolute_dist_for_single_cluster` throws an error when the cluster column is all zeros:

Error in dim(robj) <- c(dX, dY) :
  dims [product 4] do not match the length of object [1]

If the tibble is converted to a data frame first, `outer` works as expected and does not throw an error. `absolute_dist_for_single_cluster` now checks whether the input is a tibble and converts it to a data frame if it is.