Added tibble package under "Suggests" in DESCRIPTION
`absolute_dist_for_single_cluster` throws an error if the input is a tibble and the cluster fill rates are all zero. I fixed the error by converting the input to a data frame. I wrote a test to check that this fixes the error. The tibble package is required for the test but isn't used anywhere else, so I added it to DESCRIPTION as a suggested package.