Fixed bug in `get_score` by switching to ranger package.
## The problem
Many modeling packages use `stats::predict` for predictions. I don't fully understand the mechanics, but each individual modeling package includes a non-exported function that tells `stats::predict` how to work on that particular type of model. However, in order for `stats::predict` to get the necessary information, the modeling package has to be loaded with `library()`. This causes problems when trying to use `stats::predict` in another package because `library()` calls are not permitted.

## The Fix
The best work around that I could find on Google is to use the ranger package to train a random forest and call `ranger::predictions(stats::predict())` to make predictions. The randomForest package does not have a similar function, so I switched to the ranger package and implemented this solution. `devtools::check()` runs without errors, warnings, or notes.