Save densities inside the random forest instead of separately.
Moved `make_densities_from_rf` inside `train_rf`. The output of `train_rf` is now a list of the random forest, the distances data frame, and the same and different writer densities.

Added rforest as an argument for `calculate_slr`. Now rforest contains the densities as well as the random forest, the user doesn't need to input the densities to `calculate_slr` separately.

Updated data / data-raw, test fixtures, and tests.