Saved scores to random_forest list
Previously, random_forest was a list that contained (1) distances, (2) a ranger random forest object, and (3) densities fit to similarity scores.

Because we want to plot the similarity scores themselves and not the densities fit with KDE, I replaced the densities in random_forest with the scores. Now `train_rf()` returns a list of (1) distances, (2) a ranger random forest object, and (3) scores.

The densities are only used to create SLRs so I updated `calculate_slr()` to fit the densities, instead of taking the pre-fit densities as input with the random forest.

The densities are no longer available to users as a data object. If the user wants the densities, they can run `make_densities_from_rf(scores = rforest$scores)`.

The densities are saved as a test fixture and used to test `make_densities_from_rf()`.