Created `compare_writer_profiles()` to run experiments faster
Running experiments with `compare_documents()` is slow because it will process the documents and get cluster assignments everytime the function is called. `compare_writer_profiles()` has the same output as `compare_documents()` but it takes the cluster fill rates as input. This way each document only needs to be processed once an the cluster assignments only need to be calculated once.