Deleted old GitHub workflows
I made an individual GitHub workflow for R-CMD-check on Mac, Windows, and Ubuntu because the checks failed on older versions of Ubuntu but passed on Mac and Windows. I wanted badges on the README to show that at least the checks passed on Mac and Windows. Now that I fixed the problems in thw workflow that caused the checks to fail on Ubuntu, I am replacing the 3 badges with a single badge.